Fixed shaders.  Bumped version.
Former-commit-id: e2482a500840641f3e08723584db584b29032b17 [formerly 1024d6cfcd43117507d70803f9a050b952358f8f] [formerly 4967cb78c93a2b0dc22d2e146fab0ca046e75159] [formerly 7b7801d0f3bd8fae1f0c1cb89427c15df9481308] [formerly aa2e4e8fb28275621c221330be9d46045171dc4f [formerly 0cefb6b208d991738919a5bb2a28aecfe1e23e09 [formerly c8da33ae96f3a8bd3c389b331c96794b9acdbbe0] [formerly aa2e4e8fb28275621c221330be9d46045171dc4f [formerly aa2e4e8fb28275621c221330be9d46045171dc4f [formerly 1c2671420307e360c327e1bbf63021289a9a4c29]]]]] [formerly 1f5f6e2c410d71eefada26fa075d155044376bca [formerly d82137a20e4373dcb41c3a48586c87f5ddc2f228] [formerly d5c5f1f2b17ceb9056be0d2b885d7267a17470e5] [formerly b0087652a84a58a4a9895b01c8383e0785a69909] [formerly f6518083b9bc7cd2c5eeab7049411f4e47f35f02 [formerly 9645e8c04051c544c0487c90f315438b1e6449e8] [formerly b0087652a84a58a4a9895b01c8383e0785a69909 [formerly b0087652a84a58a4a9895b01c8383e0785a69909 [formerly d27051208623c25add8f3b4a3cdffc094328e236]]]] [formerly f4166e1aa6946505587729698a6ce26adc6a34f6 [formerly f4166e1aa6946505587729698a6ce26adc6a34f6 [formerly f4166e1aa6946505587729698a6ce26adc6a34f6 [formerly 7d0e58f60987076bff9cbc0ef98786c65f57e62e]]]]]
Former-commit-id: 1f5f6e2c410d71eefada26fa075d155044376bca [formerly d82137a20e4373dcb41c3a48586c87f5ddc2f228] [formerly d5c5f1f2b17ceb9056be0d2b885d7267a17470e5] [formerly b0087652a84a58a4a9895b01c8383e0785a69909] [formerly f6518083b9bc7cd2c5eeab7049411f4e47f35f02 [formerly 9645e8c04051c544c0487c90f315438b1e6449e8] [formerly b0087652a84a58a4a9895b01c8383e0785a69909 [formerly b0087652a84a58a4a9895b01c8383e0785a69909 [formerly d27051208623c25add8f3b4a3cdffc094328e236]]]]
Former-commit-id: b7a2ff561a45faa42e02a1a84c42448dc6c525db [formerly fbcf6b428a9eb7d395b1a05647c15e8d2f93e58a [formerly b4660b81e11bf88df0821160dfa23cec8eb8ebe4] [formerly b7a2ff561a45faa42e02a1a84c42448dc6c525db [formerly b7a2ff561a45faa42e02a1a84c42448dc6c525db [formerly 3d6aa2da2e4a41746b77d1556a022deb5f5ad6d4]]]]
Former-commit-id: 9adafdca57bf5b0650a28c146a84d6fb3c39b8a8
Former-commit-id: 70f1f12f0a2aace42e8bf7ce23e9bc74a9d5ca95
Former-commit-id: 9cc1a78d9ed1ef1d4c2551c16810a8477e377994
Former-commit-id: 676cb39e0684ef3508693a502482f75302f5f3df

+++ shaders/YByRy.glsl
@@ -132,17 +132,17 @@ void main()
   
   if ( mask == 1 )  // even odd rows
   {
-      float f = tc.y * height;
+      float f = float( tc.y * height );
       x = int( mod( f, 2 ) );
   }
   else if ( mask == 2 ) // even odd columns
   {
-      float f2 = tc.x * width;
+      float f2 = float( tc.x * width );
       x = int( mod( f2, 2 ) );
   }
   else if ( mask == 3 ) // checkerboard
   {
-      float f = tc.y * height;
+      float f = float( tc.y * height );
       float f2 = tc.x * width;
       x = int( mod( floor( f2 ) + floor( f ), 2 ) < 1 );
   }
